+++ src/extension.ts
@@ -6,7 +6,9 @@ import * as vscode from 'vscode';
 import {
   EXTENSION_DISPLAY_NAME,
   EXTENSION_ID,
+  EXTENSION_NAME,
   ExtensionConfig,
+  USER_PUBLISHER,
 } from './app/configs';
 import { FilesController } from './app/controllers';
 
@@ -22,7 +24,7 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.workspace.workspaceFolders.length === 0
   ) {
     const message = vscode.l10n.t(
-      'No workspace folders are open. Please open a workspace folder to use this extension'
+      'No workspace folders are open. Please open a workspace folder to use this extension',
     );
     vscode.window.showErrorMessage(message);
     return;
@@ -33,7 +35,7 @@ export async function activate(context: vscode.ExtensionContext) {
     resource = vscode.workspace.workspaceFolders[0];
   } else {
     const placeHolder = vscode.l10n.t(
-      'Select a workspace folder to use. This folder will be used to load workspace-specific configuration for the extension'
+      'Select a workspace folder to use. This folder will be used to load workspace-specific configuration for the extension',
     );
     const selectedFolder = await vscode.window.showWorkspaceFolderPick({
       placeHolder,
@@ -48,8 +50,20 @@ export async function activate(context: vscode.ExtensionContext) {
 
   // Get the configuration for the extension
   const config = new ExtensionConfig(
-    vscode.workspace.getConfiguration(EXTENSION_ID, resource?.uri)
+    vscode.workspace.getConfiguration(EXTENSION_ID, resource?.uri),
   );
+
+  // Watch for changes in the configuration
+  vscode.workspace.onDidChangeConfiguration((event) => {
+    const workspaceConfig = vscode.workspace.getConfiguration(
+      EXTENSION_ID,
+      resource?.uri,
+    );
+
+    if (event.affectsConfiguration(EXTENSION_ID, resource?.uri)) {
+      config.update(workspaceConfig);
+    }
+  });
 
   // -----------------------------------------------------------------
   // Get version of the extension
@@ -62,20 +76,50 @@ export async function activate(context: vscode.ExtensionContext) {
 
   // Check if the extension is running for the first time
   if (!previousVersion) {
-    const message = vscode.l10n.t('Welcome to {0}!', [EXTENSION_DISPLAY_NAME]);
+    const message = vscode.l10n.t(
+      'Welcome to {0} version {1}! The extension is now active',
+      [EXTENSION_DISPLAY_NAME, currentVersion],
+    );
     vscode.window.showInformationMessage(message);
 
     // Update the version in the global state
     context.globalState.update('version', currentVersion);
   }
 
-  // Check if the extension has been updated
   if (previousVersion && previousVersion !== currentVersion) {
+    // Check if the extension has been updated
+    const actions: vscode.MessageItem[] = [
+      {
+        title: vscode.l10n.t('Release Notes'),
+      },
+      {
+        title: vscode.l10n.t('Close'),
+      },
+    ];
+
     const message = vscode.l10n.t(
-      'Looks like {0} has been updated to version {1}!',
-      [EXTENSION_DISPLAY_NAME, currentVersion]
+      'New version of {0} is available. Check out the release notes for version {1}',
+      [EXTENSION_DISPLAY_NAME, currentVersion],
     );
-    vscode.window.showInformationMessage(message);
+    vscode.window.showInformationMessage(message, ...actions).then((option) => {
+      if (!option) {
+        return;
+      }
+
+      // Handle the actions
+      switch (option?.title) {
+        case actions[0].title:
+          vscode.env.openExternal(
+            vscode.Uri.parse(
+              `https://marketplace.visualstudio.com/items/${USER_PUBLISHER}.${EXTENSION_NAME}/changelog`,
+            ),
+          );
+          break;
+
+        default:
+          break;
+      }
+    });
 
     // Update the version in the global state
     context.globalState.update('version', currentVersion);
@@ -90,16 +134,16 @@ export async function activate(context: vscode.ExtensionContext) {
 
   const registerExportToFileCommand = vscode.commands.registerCommand(
     `${EXTENSION_ID}.exportToFile`,
-    (args) => filesController.exportToFile(args)
+    (args) => filesController.exportToFile(args),
   );
   const registerExportToClipboardCommand = vscode.commands.registerCommand(
     `${EXTENSION_ID}.exportToClipboard`,
-    (args) => filesController.exportToClipboard(args)
+    (args) => filesController.exportToClipboard(args),
   );
 
   context.subscriptions.push(
     registerExportToFileCommand,
-    registerExportToClipboardCommand
+    registerExportToClipboardCommand,
   );
 }